requirements.txt
google-cloud-documentai==2.9.1
google-api-core==2.11.0
pandas==1.5.3
protobuf==4.21.12
tqdm==4.64.1
